# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Recipes API
  version: 1.0.0
  description: API para gestión de recetas, ingredientes, menús, usuarios, proveedores, inventario e impuestos.
servers:
  - url: http://localhost:4000
    description: Servidor local de desarrollo
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
...
    ResponseMessage:
      type: object
      properties:
        message:
          type: string
...
    Supplier:
      type: object
      properties:
        supplier_id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        website_url:
          type: string
          format: uri
        address:
          type: string
    SupplierCreate:
      type: object
      required:
        - name
        - phone
        - email
        - website_url
        - address
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        website_url:
          type: string
          format: uri
        address:
          type: string
    SupplierIngredientItem:
      type: object
      properties:
        ingredient_id:
          type: integer
        name:
          type: string
        price:
          type: number
        delivery_time:
          type: string
          description: Tiempo de entrega (por ejemplo, "3 días")
        is_preferred_supplier:
          type: boolean
    SupplierIngredientCreate:
      type: object
      required:
        - ingredient_id
        - price
        - delivery_time
        - is_preferred_supplier
      properties:
        ingredient_id:
          type: integer
        price:
          type: number
        delivery_time:
          type: string
          description: Tiempo de entrega (por ejemplo, "3 días")
        is_preferred_supplier:
          type: boolean
    Tax:
      type: object
      properties:
        tax_id:
          type: integer
        name:
          type: string
        percent:
          type: number
          description: Porcentaje aplicado (por ejemplo, 21.0 para 21%)
    TaxCreate:
      type: object
      required:
        - name
        - percent
      properties:
        name:
          type: string
        percent:
          type: number
          description: Porcentaje aplicado (por ejemplo, 21.0 para 21%)
...
paths:
...
  /suppliers:
    get:
      tags:
        - Proveedores
      summary: Listar proveedores
      description: |
        Recupera todos los proveedores ordenados por nombre.
        **Roles permitidos**: `admin`, `supplier_manager`
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de proveedores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Supplier'

    post:
      tags:
        - Proveedores
      summary: Crear proveedor
      description: |
        Crea un nuevo proveedor en el sistema.
        **Roles permitidos**: `admin`, `supplier_manager`
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierCreate'
      responses:
        '201':
          description: Proveedor creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /suppliers/{id}:
    get:
      tags:
        - Proveedores
      summary: Obtener proveedor por ID
      description: |
        Recupera los datos de un proveedor específico.
        **Roles permitidos**: `admin`, `supplier_manager`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Datos del proveedor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
        '404':
          description: Proveedor no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Proveedores
      summary: Actualizar proveedor
      description: |
        Modifica los datos de un proveedor existente.
        **Roles permitidos**: `admin`, `supplier_manager`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierCreate'
      responses:
        '200':
          description: Proveedor actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Proveedor no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Proveedores
      summary: Eliminar proveedor
      description: |
        Elimina un proveedor y sus ingredientes asociados.
        **Roles permitidos**: `admin`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Proveedor eliminado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Proveedor no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /suppliers/{id}/ingredients:
    get:
      tags:
        - Proveedores
      summary: Listar ingredientes de proveedor
      description: |
        Obtiene los ingredientes asociados a un proveedor.
        **Roles permitidos**: `admin`, `supplier_manager`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ingredientes del proveedor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupplierIngredientItem'

    post:
      tags:
        - Proveedores
      summary: Asignar ingrediente a proveedor
      description: |
        Asocia un ingrediente a un proveedor.
        **Roles permitidos**: `admin`, `supplier_manager`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierIngredientCreate'
      responses:
        '201':
          description: Ingrediente asignado al proveedor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '400':
          description: Ingrediente ya asociado o parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /suppliers/{id}/ingredients/{ingredient_id}:
    put:
      tags:
        - Proveedores
      summary: Actualizar relación proveedor-ingrediente
      description: |
        Actualiza los datos de la relación entre un proveedor y un ingrediente (precio, tiempo de entrega, proveedor preferido).
        **Roles permitidos**: `admin`, `supplier_manager`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: ingredient_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierIngredientCreate'
      responses:
        '200':
          description: Relación actualizada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Proveedor o ingrediente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Proveedores
      summary: Desvincular ingrediente de proveedor
      description: |
        Elimina la asociación de un ingrediente con un proveedor.
        **Roles permitidos**: `admin`, `supplier_manager`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: ingredient_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ingrediente eliminado del proveedor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Ingrediente o proveedor no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /taxes:
    get:
      tags:
        - Impuestos
      summary: Listar impuestos
      description: |
        Recupera todos los impuestos ordenados por nombre.
        **Roles permitidos**: `admin`
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de impuestos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tax'

    post:
      tags:
        - Impuestos
      summary: Crear impuesto
      description: |
        Crea un nuevo impuesto.
        **Roles permitidos**: `admin`
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaxCreate'
      responses:
        '201':
          description: Impuesto creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /taxes/{id}:
    get:
      tags:
        - Impuestos
      summary: Obtener impuesto por ID
      description: |
        Recupera los datos de un impuesto específico.
        **Roles permitidos**: `admin`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Datos del impuesto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tax'
        '404':
          description: Impuesto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Impuestos
      summary: Actualizar impuesto
      description: |
        Modifica el nombre o porcentaje de un impuesto.
        **Roles permitidos**: `admin`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaxCreate'
      responses:
        '200':
          description: Impuesto actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Impuesto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Impuestos
      summary: Eliminar impuesto
      description: |
        Elimina un impuesto del sistema.
        **Roles permitidos**: `admin`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Impuesto eliminado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Impuesto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...